ns - added controllers folder and separated earthquake search and homepage controller classes. all tests pass and website runs smoothly on localhost
